+++ src/features/trading/TradingDashboard.tsx
@@ -28,6 +28,7 @@ import { PnLDashboard } from "./components/PnLDashboard";
 import { ClosedTradesHistory } from "./components/ClosedTradesHistory";
 import { SessionPerformancePanel } from "./components/SessionPerformancePanel";
 import { PerformanceMetricsPanel } from "./components/PerformanceMetricsPanel";
+import { AttributionPanel } from "./components/AttributionPanel";
 import { SignalWeightsPanel } from "./components/SignalWeightsPanel";
 import { RobinhoodChartCard } from "./components/RobinhoodChartCard";
 import { AbTestingPanel } from "./components/AbTestingPanel";
@@ -435,6 +436,7 @@ export default function TradingDashboard() {
           </TabsContent>
           <TabsContent value="history" className="space-y-6">
             <PerformanceMetricsPanel />
+            <AttributionPanel />
             <SessionPerformancePanel />
             <SignalWeightsPanel />
             <ClosedTradesHistory />